Safari-611.1.10/LayoutTests/http/tests/security/cross-frame-access-child-explicit-domain.html

<html>
<head>
    <script src="resources/cross-frame-access.js"></script>
</head>
<body>
<iframe id="aFrame" name="aFrame"></iframe>
<pre id="console"></pre>
<script>
    var url = "http://127.0.0.1:8000/security/resources/cross-frame-iframe-with-explicit-domain-set.html";
    var iframeId ="aFrame";
    var passMessage = "PASS: Cross frame access to frame explicitly setting document.domain was denied.";
    var failMessage = "Fail: Cross frame access to frame explicitly setting document.domain was allowed.";
    cannotAccessFrame(url, iframeId, passMessage, failMessage);
</script>
</body>
</html>
<html>
